requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aioredis==2.0.1
aiosignal==1.3.1
alembic==1.9.4
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.2.0
Babel==2.9.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
idna==3.4
magic-filter==1.0.9
Mako==1.2.4
MarkupSafe==2.1.2
multidict==6.0.4
python-dotenv==1.0.0
pytz==2022.7.1
SQLAlchemy==2.0.4
typing_extensions==4.5.0
uvicorn==0.20.0
yarl==1.8.2
